Update theme dropdown and switcher components
