+++ app/styles/components/_theme-dropdown.scss
@@ -2,7 +2,7 @@ $dropdown-padding: 4px;
 
 .theme-dropdown {
   position: absolute;
-  top: 110%;
+  top: 100%;
   left: 0;
   min-width: 150px;
   background: var(--background);
