+++ app/styles/components/_theme-switcher.scss
@@ -2,6 +2,7 @@
   width: 100%;
   margin-top: 8px;
   padding-top: 8px;
+  padding-bottom: 8px;
 
   border-top: 1px solid var(--gray-250);
   position: relative;
